+++ docs/components/media-and-icons/discord.tsx
@@ -2,7 +2,7 @@ import type { IconProps } from "@yamada-ui/react"
 import { Icon } from "@yamada-ui/react"
 import { forwardRef } from "react"
 
-export const Discord = forwardRef<SVGSVGElement, IconProps>(
+export const DiscordIcon = forwardRef<SVGSVGElement, IconProps>(
   ({ boxSize = "1.5em", ...rest }, ref) => {
     return (
       <Icon
@@ -22,4 +22,4 @@ export const Discord = forwardRef<SVGSVGElement, IconProps>(
   },
 )
 
-Discord.displayName = "Discord"
+DiscordIcon.displayName = "DiscordIcon"