+++ packages/components/calendar/src/use-calendar-picker.ts
@@ -16,7 +16,7 @@ import type {
 } from "react"
 import type { CalendarBaseProps, CalendarProps } from "./calendar"
 import type { UseCalendarProps } from "./use-calendar"
-import { layoutStyleProperties, useTheme } from "@yamada-ui/core"
+import { layoutStyleProperties, useI18n, useTheme } from "@yamada-ui/core"
 import {
   formControlProperties,
   useFormControlProps,
@@ -137,6 +137,7 @@ export type UseCalendarPickerProps<
 export const useCalendarPicker = <T extends UseCalendarProps<any>>(
   props: UseCalendarPickerProps<T>,
 ) => {
+  const { locale: defaultLocale } = useI18n()
   const { theme } = useTheme()
   let {
     type,
@@ -216,7 +217,7 @@ export const useCalendarPicker = <T extends UseCalendarProps<any>>(
   )
   const [containerProps, inputProps] = splitObject(rest, layoutStyleProperties)
   const {
-    isOpen,
+    isOpen: open,
     onClose,
     onOpen: onInternalOpen,
   } = useDisclosure({
@@ -230,7 +231,7 @@ export const useCalendarPicker = <T extends UseCalendarProps<any>>(
 
   const { disabled, readOnly } = formControlProps
 
-  locale ??= theme.__config?.date?.locale ?? "en"
+  locale ??= theme.__config?.date?.locale ?? defaultLocale
 
   const stringToDate = useCallback(
     (value: string): Date | undefined => {
@@ -284,16 +285,16 @@ export const useCalendarPicker = <T extends UseCalendarProps<any>>(
   }, [autoFocus, allowInput, disabled, readOnly, onInternalOpen])
 
   const onClick = useCallback(() => {
-    if (isOpen) {
+    if (open) {
       if (autoFocus && allowInput) inputRef.current?.focus()
     } else {
       onOpen()
     }
-  }, [autoFocus, allowInput, isOpen, onOpen])
+  }, [autoFocus, allowInput, open, onOpen])
 
   const onFocus = useCallback(() => {
-    if (!isOpen) onOpen()
-  }, [isOpen, onOpen])
+    if (!open) onOpen()
+  }, [open, onOpen])
 
   const onBlur = useCallback(
     (ev: FocusEvent<HTMLDivElement>) => {
@@ -303,9 +304,9 @@ export const useCalendarPicker = <T extends UseCalendarProps<any>>(
 
       if (!closeOnBlur) return
 
-      if (isOpen) onClose()
+      if (open) onClose()
     },
-    [closeOnBlur, isOpen, onClose],
+    [closeOnBlur, open, onClose],
   )
 
   const onKeyDown = useCallback(
@@ -318,7 +319,7 @@ export const useCalendarPicker = <T extends UseCalendarProps<any>>(
         [key: string]: KeyboardEventHandler<HTMLDivElement> | undefined
       } = {
         Backspace: onDelete,
-        Enter: !isOpen
+        Enter: !open
           ? (ev) => {
               ev.preventDefault()
               ev.stopPropagation()
@@ -334,7 +335,7 @@ export const useCalendarPicker = <T extends UseCalendarProps<any>>(
               onClose()
             }
           : undefined,
-        Space: !isOpen
+        Space: !open
           ? (ev) => {
               ev.preventDefault()
               ev.stopPropagation()
@@ -350,16 +351,7 @@ export const useCalendarPicker = <T extends UseCalendarProps<any>>(
 
       action(ev)
     },
-    [
-      closeOnEsc,
-      disabled,
-      readOnly,
-      isOpen,
-      onClose,
-      onEnter,
-      onDelete,
-      onOpen,
-    ],
+    [closeOnEsc, disabled, readOnly, open, onClose, onEnter, onDelete, onOpen],
   )
 
   const onClear = useCallback(
@@ -368,14 +360,14 @@ export const useCalendarPicker = <T extends UseCalendarProps<any>>(
 
       onClearProp?.()
 
-      if (autoFocus && allowInput && isOpen) inputRef.current?.focus()
+      if (autoFocus && allowInput && open) inputRef.current?.focus()
     },
-    [autoFocus, allowInput, isOpen, onClearProp],
+    [autoFocus, allowInput, open, onClearProp],
   )
 
   useOutsideClick({
     ref: containerRef,
-    enabled: isOpen && closeOnBlur,
+    enabled: open && closeOnBlur,
     handler: onClose,
   })
 
@@ -416,7 +408,7 @@ export const useCalendarPicker = <T extends UseCalendarProps<any>>(
       ...props,
       closeOnBlur: false,
       closeOnButton: false,
-      isOpen,
+      isOpen: open,
       trigger: "never",
       onClose,
       onOpen,
@@ -430,7 +422,7 @@ export const useCalendarPicker = <T extends UseCalendarProps<any>>(
       flip,
       gutter,
       isLazy,
-      isOpen,
+      open,
       lazyBehavior,
       matchWidth,
       modifiers,
@@ -448,7 +440,7 @@ export const useCalendarPicker = <T extends UseCalendarProps<any>>(
     (props = {}, ref = null) => {
       return {
         "aria-haspopup": "dialog",
-        "data-active": dataAttr(isOpen),
+        "data-active": dataAttr(open),
         "data-not-allowed": dataAttr(!readOnly && !disabled && !allowInput),
         role: "combobox",
         tabIndex: -1,
@@ -468,7 +460,7 @@ export const useCalendarPicker = <T extends UseCalendarProps<any>>(
       disabled,
       readOnly,
       formControlProps,
-      isOpen,
+      open,
       rest,
       onFocus,
       onKeyDown,
@@ -575,8 +567,10 @@ export const useCalendarPicker = <T extends UseCalendarProps<any>>(
     allowInput,
     containerRef,
     dateToString,
+    inputFormat,
     inputRef,
-    isOpen,
+    locale,
+    open,
     pattern,
     stringToDate,
     formControlProps,